feat: refactor the folder structure
