+++ jinahub/encoder/clip_image.py
